providers:resource: drop setup.{py,cfg}
This patch drops the setup.py file and the corresponding configuration
file. The idea that providers are python modules was an early convenience
where we could get a lot of features 'for free'. This approach doesn't
scale to where our users want to go (go providers, if you forget the pun)
and arguably distutils is not the best way to compile arbitrary software
in arbitrary languages.

With setup.py gone, all of the functionality needs to move to manage.py

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>