providers:resource: handle sdist specially
This patch overrides the `manage.py sdist` command to run autoreconf -i in the
src/ directory. We need to ship a number of files but they are automatically
generated and we don't want them to be a part of the version control system so
they are generated by the updated sdist command.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>