providers:resource: add packaging meta-data
This patch adds the new packaging meta-data for everything we ship in
the provider. Compared to the current Debian packaging there are some
changes. Most notably 'wodim' is not a real dependency so it has been
removed. Everything else now precisely depends on the correct package.

Some X11-related dependencies are 'Suggests' instead so that those are
not installed automatically.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>