providers:resource:src: Add some heuristic to guess 802.11ac capabilities
Broadcom drivers don't work well with libnl and 802.11ac capabilities are not
reported by the resource job.

But the name of the device tells us that ac is supported (using the lspci
output), e.g:

802.11ac Wireless Network Adapter [14e4:43b1] (rev 03)

Using the "802.11ac" pattern the resource script can now report capabilities
for broken drivers but with a meaningful name.