Fix the package resource to return at least one record
On snappy the dpkg-query is not available and the resource fails silently.
It's an issue for jobs expressing requirements as follow:

package.name == 'foo' or lsb.distributor_id == 'Ubuntu Core'

As even if UC is properly detected, the failing package resource with no
records makes the whole expression to fail (i.e == False).

This patch return a fake plainbox package name to fix the problem.